client agnostic multi-connection test
