+++ transport/transport.test.ts
@@ -1,6 +1,7 @@
 import { describe, test, expect, afterAll, vi } from 'vitest';
 import {
   createDummyTransportMessage,
+  payloadToTransportMessage,
   waitForMessage,
 } from '../util/testHelpers';
 import { EventMap } from '../transport/events';
@@ -10,6 +11,7 @@ import {
   waitFor,
 } from '../__tests__/fixtures/cleanup';
 import { testMatrix } from '../__tests__/fixtures/matrix';
+import { PartialTransportMessage } from './message';
 
 describe.each(testMatrix())(
   'transport-agnostic behaviour tests ($transport.name transport, $codec.name codec)',
@@ -226,5 +228,51 @@ describe.each(testMatrix())(
       await clientTransport.close();
       await serverTransport.close();
     });
+
+    test('multiple connections works', async () => {
+      const clientId1 = 'client1';
+      const clientId2 = 'client2';
+      const serverId = 'SERVER';
+      const serverTransport = getServerTransport();
+  
+      const makeDummyMessage = (message: string): PartialTransportMessage => {
+        return payloadToTransportMessage({ message });
+      };
+  
+      const initClient = async (id: string) => {
+        const client = getClientTransport(id);
+  
+        // client to server
+        const initMsg = makeDummyMessage('hello\nserver');
+        const initMsgId = client.send(serverId, initMsg);
+        await expect(
+          waitForMessage(serverTransport, (recv) => recv.id === initMsgId),
+        ).resolves.toStrictEqual(initMsg.payload);
+        return client;
+      };
+  
+      const client1Transport = await initClient(clientId1);
+      const client2Transport = await initClient(clientId2);
+  
+      // sending messages from server to client shouldn't leak between clients
+      const msg1 = makeDummyMessage('hello\nclient1');
+      const msg2 = makeDummyMessage('hello\nclient2');
+      const msg1Id = serverTransport.send(clientId1, msg1);
+      const msg2Id = serverTransport.send(clientId2, msg2);
+  
+      const promises = Promise.all([
+        // true means reject if we receive any message that isn't the one we are expecting
+        waitForMessage(client2Transport, (recv) => recv.id === msg2Id, true),
+        waitForMessage(client1Transport, (recv) => recv.id === msg1Id, true),
+      ]);
+      await expect(promises).resolves.toStrictEqual(
+        expect.arrayContaining([msg1.payload, msg2.payload]),
+      );
+  
+      await testFinishesCleanly({
+        clientTransports: [client1Transport, client2Transport],
+        serverTransport,
+      });
+    })
   },
 );
